+++ patzilla/util/logging/__init__.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+# (c) 2017 Andreas Motl <andreas.motl@ip-tools.org>
+from __future__ import absolute_import
+import sys
+import logging
+
+def boot_logging(options=None):
+    log_level = logging.INFO
+    if options and options.get('--debug'):
+        log_level = logging.DEBUG
+    setup_logging(level=log_level)
+
+def setup_logging(level=logging.INFO):
+    log_format = '%(asctime)-15s [%(name)-20s] %(levelname)-7s: %(message)s'
+    logging.basicConfig(
+        format=log_format,
+        stream=sys.stderr,
+        level=level)
+